requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.30.0
notebooklm-py>=0.3.0
python-multipart>=0.0.6
alembic>=1.13.0
